fix: split general tests by stack Created new tests files for each stack on the folder tests
Resolves #142
